move hackathoncard into its own file
